btrfs-progs: receive: escape target paths for rename, symlink and hardlink
The stream dump escapes the path on which the operation is done but
there are a some that use another path that's the target. A file with
eg. a newline then does not format properly on one line as expected.

Extend the printing helpers to skip printing the newline and then print
the escaped path manually.

Issue: #510
Signed-off-by: David Sterba <dsterba@suse.com>